refactor(api): extract attestation-storage mocks from test-helper.js
